Extract #meta_description into MetaDescription concern
