+++ app/models/book.rb
@@ -55,22 +55,6 @@ class Book < ApplicationRecord
     [asset_base_url_prefix, filename].join('/')
   end
 
-  def meta_description
-    if summary.blank?
-      html = Kramdown::Document.new(
-        content,
-        input: :kramdown,
-        remove_block_html_tags: false,
-        transliterated_header_ids: true
-      ).to_html.to_s
-
-      doc = Nokogiri::HTML(html)
-      doc.css('body').text.truncate(200)
-    else
-      summary
-    end
-  end
-
   # Overrides
   def ask_for_donation?
     downloads_available?
