+++ app/models/poster.rb
@@ -56,22 +56,6 @@ class Poster < ApplicationRecord
     [asset_base_url_prefix, filename].join('/')
   end
 
-  def meta_description
-    if summary.blank?
-      html = Kramdown::Document.new(
-        content,
-        input: :kramdown,
-        remove_block_html_tags: false,
-        transliterated_header_ids: true
-      ).to_html.to_s
-
-      doc = Nokogiri::HTML(html)
-      doc.css('body').text.truncate(200)
-    else
-      summary
-    end
-  end
-
   def meta_image
     front_image
   end
