+++ app/models/zine.rb
@@ -54,20 +54,4 @@ class Zine < ApplicationRecord
     filename = filename.join
     [asset_base_url_prefix, filename].join('/')
   end
-
-  def meta_description
-    if summary.blank?
-      html = Kramdown::Document.new(
-        content,
-        input: :kramdown,
-        remove_block_html_tags: false,
-        transliterated_header_ids: true
-      ).to_html.to_s
-
-      doc = Nokogiri::HTML(html)
-      doc.css('body').text.truncate(200)
-    else
-      summary
-    end
-  end
 end
